refactor out parseCsv
